feat: Add new message type for object destruction
Add a new message type, `ObjectDestroyMessage`, to the server.
This message is used to signal the destruction of an object to
other clients. It includes the network ID of the object being
destroyed.

This change also updates the `on_data` function in the server to
handle the new message type. When the server receives an
`ObjectDestroyMessage`, it will broadcast the message to all
other clients.

This change also updates the `disconnect` function in the server to
send an `ObjectDestroyMessage` for each object owned by the
disconnecting client before removing the client from the
connection map.

This change also updates the `Server` struct to use a
`HashMap<u64, Connection>` instead of a `HashMap<u64,
Connection>`. This is done to improve the performance of the
server by reducing the number of hash table lookups required to
access connections.