feat: add EntityStateMessage
Adds EntityStateMessage and sends it to observers.

This commit adds a new message type, EntityStateMessage, which is used to send the state of an entity to its observers. This allows the server to update the state of entities on the clients in real time.

The following changes were made to implement this feature:

- Added EntityStateMessage to the messages module.
- Modified the NetworkBehaviour trait to include a method for sending an EntityStateMessage.
- Updated the NetworkServer to send EntityStateMessages to observers when the state of an entity changes.
- Updated the NetworkTransformUnreliable component to use EntityStateMessages to update the interpolation and broadcast states of entities.
- Updated the PlayerScript to use EntityStateMessages to send the player's state to observers.