refactor: Use NetworkBehaviour map for command invocation
Refactor network animator to use the global
`NETWORK_BEHAVIOURS` map to fetch components
during command invocation, instead of relying on
the `NetworkIdentity`. This change enables more
efficient and flexible access to behaviours by
using the network id and component index. Error
logging is also added if the behaviour is not
found or locked.